requirements.txt
langchain==0.0.232
python-dotenv~=1.0.0
uvicorn~=0.22.0
fastapi~=0.100.0
requests~=2.31.0
pydantic~=1.10.11
openai~=0.27.8